Dipes Text Report: Exclude those who have no pay
If you don't have vacation pay, or regular pay, you
should be excluded from the report.

Note: I also added a small test to execute the report
which will fail if the SQL cannot be run. This should
be added for all the other tests as a sanity check.

Actually checking report output would be another thing
that should be added.