Supplemental Vacation Simplification.
=Overview=

This commit modifies the supplemental days behavior in payslips so that
any supplemental days are added to vacation_earned each period as they
are earned instead of building up a bank of supplemental days that needs
to be transferred periodically. This removes some complexity from the
system.

Thus, each month the number of vacation days earned will be variable.
Previously it had always been 1.5 except in the months where a transfer
occured.

Also note that the `vacation_pay_balance` in the payslip is just the
employee's vacation daily rate multiplied by employee's vacation balance.
It is recomputed each month and is no longer a running balance. Based
on conversations with finance, this is the desired behavior. The behaviors
of `vacation_used` and `vacation_pay_used` are unchanged.

`vacation_pay_earned` is still 1/16th of taxable pay, but isn't added to
any bank. This may not be a correct anymore given that disconnect.

Also adjust tests as necessary, remove methods and classes that are no
longer necessary, and perform various clean up, etc.

This also adds a test for the migration behavior.

=Migration=

An important part of this change is migrating the system's data from the
"old way" to the "new way" of doing supplemental days. This will be done
as part of the core rails migration system.

Prior to the migration to delete the suppl fields, and additional db
migration has been added that will take any existing supplemental days will
in the current payslip and create a payslip correction with the amount in
their supplemental days bank. This amount is also added to the vacation
balance immediately. The payslip corrections are needed to ensure that the
additional balance is kept when the payslip is reprocessed, otherwise there
would be no history of the addition days.

Vacation pay will be left alone since a reprocess will recompute the vacation
pay balance total.