Enter All Hours
Provide a button to enter default hours for all who need them. The
option to enter the hours employee by employee is still there. A
new function has been added to allow for filling default hours in
the model.

This should save time at the beginning of the month by allowing
a baseline to be entered and then modifications can be made to
the default hours as appropriate. This was how finance was using
the system anyways.