Generate Timesheets for Selected Employees
This changes allows all, one, or some employees to be selected for
generataion in the timesheet PDF file for printing and distributing
to employees. This change adds a multi-select on the timesheet page
with all currently paid employees selected by default.

Previously only a PDF for all employees could be generated.