Update report, include employee is payslip exists
This commit changes the behavior of most of the reports.
Previously the reports would look at a number of tables,
but also the employee's employment status to decide what
to include in the report.

The issue is that the employee's employment status is a
real-time value that is not tied to its state in a
previous pay period.

Thus, if I changed an employee's employments status today,
it had the possible side effect of removing this employee
from a previous period's reports.

This fixes this behavior to look at the existence of a
payslip to decide if the employee should be included or
not in a given period. If the employee should not be
included, then there is no payslip generated.