Create zero'd payslip if the employee is on leave
When an employee is on leave, they still need to appear in
certain reports so that certain declarations can be made
with appropriate agencies. However, the employee is also not
paid during this period.

Thus, we generate a blank payslip (similar to when an employee
is on vacation and does not work) so they will appear in the
appropriate reports, even if the pay is zero.