Update Work Loans
1. Allow work loans greater than 8 hours. This was cumbersome
to enter each day as a separate work loan. You can still do this
but now you can enter 1 work loan per month with the total hours
for the month and everything will still work fine. (i.e. one
work loan on 9/1 for 115 hours). This is simpler to enter.

2. Updated percentage calculation to use actual hours worked per
month. Previously, if you worked a full month, it used the
"standard hours" worked per month. This was pretty close, but
I think the calculation is better now.