[Department Charge Report] Do not duplicate count
Employees may appear on this report multiple times if their work
is loaned to another department. In this case, part of their
salary will be charged to multiple departments. This works fine,
but the employee's records will appear multiple times in the report
inflating other records. Change this to not add duplicate records
to the total for employees that have already appeared on the report.