Some Employees do not Accrue Vacation
Employees can now have their vacation accural toggled
in their employee record. This can be enabled for a
period, and then enabled again wherein the system will
start counting again normally. The system will report
zeros during the time it is off and display of the
vacation balance will be suppressed.

Note: This commit also adds a hint that you are on dev
in bright red. Because prod and dev look the same you
can get confused.