Fix issue with only vacation payslip
If an employee only had a vacation payslip and not a regular
payslip in that month, the user's pay information did not
appear in the DIPES reports. This led to an imbalance in
other reports which used those reports.

This changes the queries so that this situation is handled
and will appear in the reports properly.