Exceptional Raises vs Normal Raises
The system will alert you when an echelon raise is due.
This is normally two years after the last echelon raise.

However, there are some raises which are provided outside
of the normal two-year schedule. These are exceptional
raises and do not reset the two year time for the reminder.

The reminder will be shown when two years passes since the
last non-exceptional (or "normal") raise.